rename `formatPayload` → `formatSignupPayload`
